+++ EliteJournalReader/Events/SupercruiseExitEvent.cs
@@ -19,10 +19,10 @@ namespace EliteJournalReader.Events
 
         public class SupercruiseExitEventArgs : JournalEventArgs
         {
-            public long  SystemAddress { get; set; }
+            public long SystemAddress { get; set; }
             public string StarSystem { get; set; }
             public string Body { get; set; }
-            public long BodyID { get; set; }
+            public int BodyID { get; set; }
 
             [JsonConverter(typeof(ExtendedStringEnumConverter<BodyType>))]
             public BodyType BodyType { get; set; }
